probin for pltsql func/proc is not restored properly by pg_restore
Since PG does not store typmod for function/procedure’s
input and output types, we store that information in probin
for T-SQL function/procedure. However, it is not properly
recovered by pg_restore.

The reason for this is that we intentionally ignore
probin string in pltsql_function_as_checker function
and then try to generate a new probin later while creating
the function/procedure.

Fixed this by taking probin string from AS clause
if it is provided and skipping generating the new
probin during function/procedure creation.

Signed-off-by: Rishabh Tanwar <ritanwar@amazon.com>