Fix MVU failure due to a function removed from a new binary
Revive drop_all_users() with an empty body as it is needed during MVU.
More specifically, we cannot remove this function entirely because, in
PG13, sys.babel_drop_all_users() procedure refers it.

Removing the procedure sys.babel_drop_all_users() during pg_dump cannot
be an option because other user-defined procedures are able to refer
this function as well.

Signed-off-by: Jungkook Lee <jungkook@amazon.com>