OPENQUERY for Linked Servers does not work when used from user created database (#1592)
Description
Previously the user mapping that got created as part of sp_addlinkedsrvlogin was only working in master database. In T-SQL the linked login is not specific to login unlike the user mappings in PG. To fix this we create the user mapping for the public role so that the linked login is not specific to any single role.

Task: BABEL-4229

Signed-off-by: Sharu Goel goelshar@amazon.com