Fix the issue that inserted table refered in update join will cause crash (#2140)
Previously when we implement update join, we missed a corner case that
one of the join relations can be a named tuple store instead of RTE
relation. And it'll lead analyzer to wrongly set resultRelation for the
Query.
This fix resolved the corner case by add a if condition to exclude
RTE_NAMEDTUPLESTORE for being resultRelation.

Task: BABEL-4606
Signed-off-by: Zhibai Song <szh@amazon.com>