Additional fixes on top of commit for introducing bbf_role_admin role
PG community has completely overhauled the permissions system for
PG16 roles. One of the change is that creator of a role gets added
as a member of the role with ADMIN OPTION.

Previously, we would default to using sysadmin as the creator when
creating/modifying/deleting roles, but this no longer works since
sysadmin itself is a role, and roles get automatically added with
ADMIN OPTION to new roles which they create, then the newly created
role can no longer be added into the sysadmin role (since that
would cause a circular dependency).

This commit fixes this by indroducing a new Babelfidh role called
bbf_role_admin which will be used for creating database objects
instead of sysadmin.

Signed-off-by: Rishabh Tanwar <ritanwar@amazon.com>