fix for DB user drop query when granted some permission (#3789) (#4088)
Problem : If a user having some permissions over an object then we are not able to drop the user, which is not same as TSQL behaviour.Note: in PostgreSQL we can't drop a role before dropping it's dependencies.
Fix: we have to remove the grants(permission) given to the user via shared_dependency_acl before dropping the user.

    Added a check if user is grantor of some permissions then it can't be dropped.
    After dropping the user remove the entries corresponding to that user from schema permission catalog.
    To implement we have created a new function bbf_shdep_drop_owned_dependent_acl (forked from pg function shdepDropOwned).

Task: BABEL-3026

Signed-off-by: pranav jain <pranav23iitd@gmail.com>