Fix crash when pg_cron background jobs throw errors in Babelfish (#4127)
Problem: When pg_cron is enabled with babelfish and we run a background job which is obvious to throw some error, then we are landing into a crash

RCA: The problem is when we run a pg_cron and execute a query via background worker then in terminate_batch function memory context MessageContext is not initialised( or is NULL), which leads to crash.
Fix: Check if we are connected via background worker then do not switch to MessageContext instead, throw error afterwards.

Issues Resolved
[BABEL-5995]

Signed-off-by: pranav jain <pranav23iitd@gmail.com>

---------

Signed-off-by: pranav jain <pranav23iitd@gmail.com>